Seperate listener / handler for file events
